Laid foundation
Been coding for 8 hours.
My eyes.
The pain.
Not looking forward to looking at this code again.

I now understand what I need to know for this test question tomorrow

To future me: Didn't do frequent commits bc I was 2 files in when I decided to create the repository (and git decided to create the repository in a different folder)